requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15
django-cors-headers>=4.0.0,<4.1
django-extensions>=3.2.1,<3.3
django-jazzmin>=2.6.0,<2.7
psycopg[c]>=3.1.8,<3.2
gunicorn>=20.1.0,<20.2
djangorestframework-simplejwt>=5.2.2,<5.3
django-silk>=5.0.3,<5.1
cryptography>=41.0.0,<=41.0.5
